requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.0.0
anthropic>=0.7.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
transformers>=4.35.0
torch>=2.1.0
pandas>=2.0.0